CS210_2016 & 2017 solutions
The lab questions are used for the final exams and they are slight variations on eachother.  I recommend looking up the previous exam papers, as the questions are sometimes taken directly from them with minor alterations.

For a solid understanding of the data structures, and for an understanding of the course in general, the book Data Structures and Algorithms is your best friend.  It's an easy read and does a solid job explaining how the code works, along with the concepts behind them.

The labs are stressful, unnecessarily so in my opinion.  However, the code provided here will help you get through it and go on to do well.

One last tip, Phil hints at what one questions are coming up.  It is worth recording the lectures, especially the final two lectures where he tells you what sort of questions are coming up.  When you are listening and taking down notes, you are only taking in approximately 10% of what is said.  With a recording you will take in information and clues even after five times listening to it.

Best of luck and enjoy it if you can